feat: move profile card to hero section, remove contact section
