+++ src/App.tsx
@@ -1,6 +1,5 @@
 import About from './components/About/About';
 import Achievements from './components/Achievements/Achievements';
-import Contact from './components/Contact/Contact';
 import Experience from './components/Experience/Experience';
 import Copyright from './components/Footer/Footer';
 import NavBar from './components/NavBar/NavBar';
@@ -21,7 +20,6 @@ const App = () => {
             <Experience />
             <Achievements />
             <Projects />
-            <Contact />
           </div>
         </div>
       </main>
